+++ core/src/main/java/io/questdb/cairo/idx/PostingIndexBwdReader.java
@@ -335,7 +335,7 @@ public class PostingIndexBwdReader extends AbstractPostingIndexReader {
 
         private void loadDenseGenerationCached(int gen) {
             long genFileOffset = genLookup.getGenFileOffset(gen);
-            int genDataSize = genLookup.getGenDataSize(gen);
+            long genDataSize = genLookup.getGenDataSize(gen);
             int genKeyCount = genLookup.getGenKeyCount(gen);
 
             if (requestedKey >= genKeyCount) {
@@ -454,7 +454,7 @@ public class PostingIndexBwdReader extends AbstractPostingIndexReader {
          */
         private void loadSparseGenDirect(int gen, int idx) {
             long genFileOffset = genLookup.getGenFileOffset(gen);
-            int genDataSize = genLookup.getGenDataSize(gen);
+            long genDataSize = genLookup.getGenDataSize(gen);
             int genKeyCount = genLookup.getGenKeyCount(gen);
             int activeKeyCount = -genKeyCount;
 
@@ -482,7 +482,7 @@ public class PostingIndexBwdReader extends AbstractPostingIndexReader {
 
         private void loadSparseGenWithBinarySearch(int gen) {
             long genFileOffset = genLookup.getGenFileOffset(gen);
-            int genDataSize = genLookup.getGenDataSize(gen);
+            long genDataSize = genLookup.getGenDataSize(gen);
             int genKeyCount = genLookup.getGenKeyCount(gen);
             int activeKeyCount = -genKeyCount;
